refactoring. carry common functions to util
