+++ src/api/github.ts
@@ -9,6 +9,7 @@
  * Holds important api calls for the crates.io.
  */
 import { get } from "request-promise";
+import { decidePath, parseVersions } from "./index-utils";
 
 const API = "https://api.github.com/repos/rust-lang/crates.io-index";
 
@@ -19,7 +20,7 @@ function cache(key: string, func: any, url: string, githubToken?: string) {
 
   if (!data[key] || data[key].isRejected()) {
     console.log("Fetching dependency: ", key);
-    const headers: { [key: string]: string } = {
+    const headers: { [key: string]: string; } = {
       "User-Agent":
         "VSCode.Crates (https://marketplace.visualstudio.com/items?itemName=serayuzgur.crates)",
       Accept: "application/vnd.github.VERSION.raw",
@@ -30,22 +31,7 @@ function cache(key: string, func: any, url: string, githubToken?: string) {
     data[key] = func(url, {
       headers
     })
-      .then((response: string) => {
-        const conv = response.split("\n");
-        console.log("Fetching DONE: ", key, conv.length);
-        const versions = [];
-        for (const rec of conv) {
-          try {
-            if (rec.trim().length > 0) {
-              const parsed = JSON.parse(rec);
-              versions.push({ num: parsed.vers, yanked: parsed.yanked });
-            }
-          } catch (er) {
-            console.log(er, rec);
-          }
-        }
-        return { versions: versions.sort().reverse() };
-      })
+      .then((response: string) => parseVersions(response, key))
       .catch((resp: any) => {
         console.error(resp);
         throw resp;
@@ -57,21 +43,3 @@ function cache(key: string, func: any, url: string, githubToken?: string) {
 export const versions = (name: string, githubToken?: string) => {
   return cache(name, get, `${API}/contents/${decidePath(name)}`, githubToken);
 };
-
-export function decidePath(name: string) {
-  name = name.toLowerCase();
-  if (name.startsWith('"') && name.endsWith('"')) {
-    name = name.substring(1, name.length - 1);
-  }
-  if (name.length === 1) {
-    return `1/${name}`;
-  }
-  if (name.length === 2) {
-    return `2/${name}`;
-  }
-  if (name.length === 3) {
-    return `3/${name.charAt(0)}/${name}`;
-  }
-
-  return `${name.substring(0, 2)}/${name.substring(2, 4)}/${name}`;
-}
